pyproject.toml
[tool.poetry]
name = "dan-toropov-hw-7"
version = "0.1.0"
description = ""
authors = ["dan.toropov <dan.toropov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
pillow = "^11.0.0"
psycopg-binary = "^3.2.3"
python-dotenv = "^1.0.1"
psycopg = "^3.2.3"
djangorestframework = "^3.15.2"
black = "^24.10.0"
isort = "^5.13.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
